Виправляє логіку відкриття модалки, виносить розмітку в окремий файл
