+++ src/sass/base/_common.scss
@@ -1,31 +1,20 @@
 @import '~node_modules/modern-normalize/modern-normalize.css';
 
 body {
-    font-style: normal;
-    font-family: 'Roboto', sans-serif;
-    font-weight: 500;
-    @include lh(16px, 12px, 12px);
-    background-color: getColor(white-text-color);
-    color: getColor(primary-text-color);
+  font-style: normal;
+  font-family: 'Roboto', sans-serif;
+  font-weight: 500;
+  @include lh(16px, 12px, 12px);
+  background-color: getColor(white-text-color);
+  color: getColor(primary-text-color);
 
-    @media screen and (min-width: getSize(desktop)) {
-        @include lh(24px, 20px, 20px);
-    }
-}
+  @media screen and (min-width: getSize(desktop)) {
+    @include lh(24px, 20px, 20px);
+  }
 
-// body {
-
-
-//     &.modal-open,
-//     &.is-open {
-//         overflow: hidden;
-//     }
-
-// }
-
-body {
-
-    &.js-modal-is-hidden {
-        overflow: hidden;
-    }
+  &.modal-open,
+  &.is-open,
+  &.js-modal-is-hidden {
+    overflow: hidden;
+  }
 }
